requirements.txt
Django==1.11.21
django_filter==1.1.0
djangorestframework==3.9.4
Faker==3.0.1
Django==1.11.21
django-cors-headers==3.0.2
django-filter==1.1.0
django-model-utils==3.1.2
djangorestframework==3.9.4
fake-factory==9999.9.9
Faker==3.0.1
ipaddress==1.0.23
Markdown==3.1.1
python-dateutil==2.8.1
pytz==2019.1
six==1.14.0
text-unidecode==1.3
